requirements/base.txt
anndata~=0.8.0
fastavro
google-api-python-client
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.9.0
google-cloud-bigquery-storage==2.19.1
google-cloud-storage
scipy
fastapi==0.96.0
uvicorn[standard]
python-dotenv
pyarrow
python-multipart
aiohttp==3.8.3
SQLAlchemy==1.4.46
Flask==2.2.2
Flask-Admin==1.6.0
Flask-HTTPAuth==4.7.0
PyJWT==2.6.0
alembic==1.9.2
gunicorn==20.1.0
pg8000==1.29.4
cloud-sql-python-connector==1.1.0
neptune==1.1.1
matplotlib
pandas
google-cloud-run==0.7.1
scikit-learn
lightning==2.2.0
db-dtypes==1.1.1
Mako==1.2.4
smart-open==6.4.0
sqlparse==0.4.4
Werkzeug==2.2.2
sentry-sdk==1.39.2
typer==0.9.0
PyYAML==6.0.1
requests==2.31.0